:building_construction: changin API for jest-mock-apollo to use factory
