+++ packages/jest-mock-apollo/src/test/index.test.tsx
@@ -2,19 +2,26 @@ import * as React from 'react';
 import {graphql} from '@shopify/react-apollo';
 
 import {mount} from 'enzyme';
+import {readFileSync} from 'fs';
+import * as path from 'path';
 
-import {registerSchema, createGraphQLClient} from '..';
+import GraphQLClientFactory, {GraphQLMock, GraphQLClientOptions} from '..';
 import unionOrIntersectionTypes from './fixtures/schema-unions-and-interfaces.json';
 import petQuery from './fixtures/PetQuery.graphql';
 
 // setup
-registerSchema({
-  schemaBuildPath: './fixtures/schema.graphql',
+const graphQLClientFactory = new GraphQLClientFactory({
+  schemaSrc: readFileSync(
+    path.resolve(__dirname, './fixtures/schema.graphql'),
+    'utf8',
+  ),
   unionOrIntersectionTypes,
 });
 
+const createGraphQLClient = graphQLClientFactory.create;
+
 interface Props {
-  data: {
+  data?: {
     loading: boolean;
   };
 }
@@ -24,10 +31,10 @@ function SomePageBase({data: {loading}}: Props) {
   return <p>{loading ? 'Loading' : 'Loaded!'}</p>;
 }
 
-const SomePage = graphql(petQuery)(SomePageBase) as React.ComponentType;
+const SomePage = graphql(petQuery)(SomePageBase);
 
 describe('jest-mock-apollo', () => {
-  it('just works', () => {
+  it('provides the required context', () => {
     mount(<SomePage />, {
       context: {
         client: createGraphQLClient(),
